[Refactor] `InsightQueryDomainService`, `CommandDomainService` 분리
