+++ keewe-api/src/main/java/ccc/keeweapi/service/insight/command/InsightCommandApiService.java
@@ -5,7 +5,8 @@ import ccc.keeweapi.dto.insight.*;
 import ccc.keeweapi.utils.annotations.TitleEventPublish;
 import ccc.keewecore.consts.TitleCategory;
 import ccc.keewedomain.persistence.domain.insight.Insight;
-import ccc.keewedomain.service.insight.InsightDomainService;
+import ccc.keewedomain.service.insight.command.BookmarkCommandDomainService;
+import ccc.keewedomain.service.insight.command.InsightCommandDomainService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,28 +15,29 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class InsightCommandApiService {
 
-    private final InsightDomainService insightDomainService;
+    private final BookmarkCommandDomainService bookmarkCommandDomainService;
+    private final InsightCommandDomainService insightCommandDomainService;
     private final InsightAssembler insightAssembler;
 
     @Transactional
     @TitleEventPublish(titleCategory = TitleCategory.INSIGHT)
     public InsightCreateResponse create(InsightCreateRequest request) {
-        Insight insight = insightDomainService.create(insightAssembler.toInsightCreateDto(request));
+        Insight insight = insightCommandDomainService.create(insightAssembler.toInsightCreateDto(request));
         return insightAssembler.toInsightCreateResponse(insight);
     }
 
     public InsightViewIncrementResponse incrementViewCount(Long insightId) {
-        Long viewCount = insightDomainService.incrementViewCount(insightAssembler.toInsightViewIncrementDto(insightId));
+        Long viewCount = insightCommandDomainService.incrementViewCount(insightAssembler.toInsightViewIncrementDto(insightId));
         return insightAssembler.toInsightViewIncrementResponse(viewCount);
     }
 
     public BookmarkToggleResponse toggleInsightBookmark(Long insightId) {
-        boolean isBookmark = insightDomainService.toggleInsightBookmark(insightAssembler.toBookmarkToggleDto(insightId));
+        boolean isBookmark = bookmarkCommandDomainService.toggleInsightBookmark(insightAssembler.toBookmarkToggleDto(insightId));
         return insightAssembler.toBookmarkToggleResponse(isBookmark);
     }
 
     @Transactional
     public InsightDeleteResponse delete(Long id) {
-        return InsightDeleteResponse.of(insightDomainService.delete(insightAssembler.toInsightDeleteDto(id)));
+        return InsightDeleteResponse.of(insightCommandDomainService.delete(insightAssembler.toInsightDeleteDto(id)));
     }
 }
