+++ keewe-api/src/main/java/ccc/keeweapi/service/insight/query/InsightQueryApiService.java
@@ -2,13 +2,18 @@ package ccc.keeweapi.service.insight.query;
 
 import ccc.keeweapi.component.InsightAssembler;
 import ccc.keeweapi.component.ProfileAssembler;
-import ccc.keeweapi.dto.insight.*;
+import ccc.keeweapi.dto.insight.ChallengeRecordResponse;
+import ccc.keeweapi.dto.insight.InsightAuthorAreaResponse;
+import ccc.keeweapi.dto.insight.InsightGetForHomeResponse;
+import ccc.keeweapi.dto.insight.InsightGetResponse;
+import ccc.keeweapi.dto.insight.InsightMyPageResponse;
 import ccc.keeweapi.utils.SecurityUtil;
 import ccc.keewedomain.dto.insight.InsightGetDto;
 import ccc.keewedomain.persistence.domain.challenge.ChallengeParticipation;
 import ccc.keewedomain.persistence.domain.insight.Insight;
 import ccc.keewedomain.persistence.repository.utils.CursorPageable;
-import ccc.keewedomain.service.insight.InsightDomainService;
+import ccc.keewedomain.service.insight.command.InsightCommandDomainService;
+import ccc.keewedomain.service.insight.query.InsightQueryDomainService;
 import ccc.keewedomain.service.user.ProfileDomainService;
 import java.util.List;
 import java.util.Objects;
@@ -21,20 +26,20 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class InsightQueryApiService {
 
-    private final InsightDomainService insightDomainService;
+    private final InsightQueryDomainService insightQueryDomainService;
     private final ProfileDomainService profileDomainService;
     private final InsightAssembler insightAssembler;
     private final ProfileAssembler profileAssembler;
 
     @Transactional(readOnly = true)
     public InsightGetResponse getInsight(Long insightId) {
-        InsightGetDto insightGetDto = insightDomainService.getInsight(insightAssembler.toInsightDetailDto(insightId));
+        InsightGetDto insightGetDto = insightQueryDomainService.getInsight(insightAssembler.toInsightDetailDto(insightId));
         return insightAssembler.toInsightGetResponse(insightGetDto);
     }
 
     @Transactional(readOnly = true)
     public InsightAuthorAreaResponse getInsightAuthorAreaInfo(Long insightId) {
-        Insight insight = insightDomainService.getByIdWithWriter(insightId);
+        Insight insight = insightQueryDomainService.getByIdWithWriter(insightId);
         boolean isFollowing = profileDomainService.getFollowingTargetIdSet(profileAssembler.toFollowCheckDto(insight.getWriter().getId()));
         return insightAssembler.toInsightAuthorAreaResponse(insight, isFollowing);
     }
@@ -42,13 +47,13 @@ public class InsightQueryApiService {
 
     @Transactional(readOnly = true)
     public List<InsightGetForHomeResponse> getInsightsForHome(CursorPageable<Long> cPage, Boolean follow) {
-        return insightDomainService.getInsightsForHome(SecurityUtil.getUser(), cPage, follow).stream()
+        return insightQueryDomainService.getInsightsForHome(SecurityUtil.getUser(), cPage, follow).stream()
                 .map(insightAssembler::toInsightGetForHomeResponse)
                 .collect(Collectors.toList());
     }
 
     public ChallengeRecordResponse getChallengeRecord(Long insightId) {
-        Insight insight = insightDomainService.getByIdWithChallengeOrElseThrow(insightId);
+        Insight insight = insightQueryDomainService.getByIdWithChallengeOrElseThrow(insightId);
         ChallengeParticipation participation = insight.getChallengeParticipation();
 
         if (Objects.isNull(participation)) {
@@ -56,7 +61,7 @@ public class InsightQueryApiService {
         }
 
         if (insight.isValid()) {
-            Long order = insightDomainService.getRecordOrder(participation, insightId);
+            Long order = insightQueryDomainService.getRecordOrder(participation, insightId);
             return insightAssembler.toChallengeRecordResponse(participation, order);
         }
 
@@ -65,7 +70,7 @@ public class InsightQueryApiService {
 
     @Transactional(readOnly = true)
     public List<InsightMyPageResponse> getInsightsForMyPage(Long userId, Long drawerId, CursorPageable<Long> cPage) {
-        return insightDomainService.getInsightsForMyPage(SecurityUtil.getUser(), userId, drawerId, cPage).stream()
+        return insightQueryDomainService.getInsightsForMyPage(SecurityUtil.getUser(), userId, drawerId, cPage).stream()
                 .map(insightAssembler::toInsightMyPageResponse)
                 .collect(Collectors.toList());
     }
