+++ keewe-statistics/src/main/java/ccc/keewestatistics/listener/InsightViewListener.java
@@ -1,8 +1,9 @@
 package ccc.keewestatistics.listener;
 
 import ccc.keewecore.consts.KeeweConsts;
-import ccc.keewedomain.service.insight.InsightDomainService;
+import ccc.keewedomain.service.insight.command.InsightCommandDomainService;
 import com.rabbitmq.client.Channel;
+import java.io.IOException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.core.Message;
@@ -11,27 +12,23 @@ import org.springframework.amqp.support.AmqpHeaders;
 import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.stereotype.Component;
 
-import java.io.IOException;
-
 @Component
 @Slf4j
 @RequiredArgsConstructor
 public class InsightViewListener {
-    private final InsightDomainService insightDomainService;
+    private final InsightCommandDomainService insightCommandDomainService;
 
     @RabbitListener(queues = KeeweConsts.INSIGHT_VIEW_QUEUE, ackMode = "MANUAL")
     public void onMessage(Message message, Channel channel, @Header(AmqpHeaders.DELIVERY_TAG) long tag) throws IOException {
-
         try {
             Long insightId = Long.parseLong(new String(message.getBody()));
 
             log.info("[IVL::onMessage] insightId {}", insightId);
-            insightDomainService.incrementViewCount(insightId);
+            insightCommandDomainService.incrementViewCount(insightId);
             channel.basicAck(tag, false);
         } catch (Throwable t) {
             log.error(t.getMessage(), t);
             channel.basicNack(tag, false, false);
         }
-
     }
 }
